Rename: add directory auth group by auth middleware
