+++ src/middleware/auth/managerAuth.js
@@ -1,7 +1,7 @@
-const pool = require("../../config/database/postgresql");
-const validate = require("../module/validation");
-const { position } = require("../module/global");
-const { ForbbidenException } = require('../module/customError');
+const pool = require("../../../config/database/postgresql");
+const validate = require("../../module/validation");
+const { position } = require("../../module/global");
+const { ForbbidenException } = require('../../module/customError');
 
 module.exports = async (req, res, next) => {
     const userId = req.decoded.id;
