+++ src/routes/account.js
@@ -1,6 +1,6 @@
-const pool = require("../../config/database/postgresql");
 const router = require("express").Router();
-const loginAuth = require("../middleware/loginAuth");
+const pool = require("../../config/database/postgresql");
+const loginAuth = require("../middleware/auth/loginAuth");
 const validate = require("../module/validation");
 const bcryptUtil = require("../module/bcrypt");
 const personalColor = require("../module/personalColor");
