+++ src/routes/notice.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
 const pool = require("../../config/database/postgresql");
-const loginAuth = require("../middleware/loginAuth");
+const loginAuth = require("../middleware/auth/loginAuth");
 const validate = require("../module/validation");
 const { club } = require("../module/global");
 const { BadRequestException } = require("../module/customError");
